Updated Files
I added a bunch of more characters, strokes and grammar topics to the json files and md. I also wanted this to be containerized so that people on different OS can use so I added docker. This new version has not been tested yet. Used JMDict project sourcing as well. Working on drawing agentic flow and testing to make sure the code can be deployed.